+++ _meta.lua
@@ -7,6 +7,6 @@ return {
     name = "updatesmanager",
     fullname = _("Updates Manager"),
     description = _([[Manage updates for patches and plugins GitHub repositories.]]),
-    version = "1.3.4",
+    version = "1.3.5",
 }